+++ pkg/clustersync/clusterSync.go
@@ -5,8 +5,19 @@ package clustersync
 import (
 	"context"
 
+	"github.com/stolostron/search-indexer/pkg/config"
 	klog "k8s.io/klog/v2"
 )
+
+func ElectLeaderAndStart(ctx context.Context) {
+	client := config.Cfg.KubeClient
+	lockName := "search-indexer.open-cluster-management.io"
+	podName := config.Cfg.PodName
+	podNamespace := config.Cfg.PodNamespace
+
+	lock := getNewLock(client, lockName, podName, podNamespace)
+	runLeaderElection(ctx, lock)
+}
 
 func syncClusters(ctx context.Context) {
 	klog.Info("TODO: Start Sync clusters here.")
